refactor(app): collapse query cell factory into single component
